requirements.txt
phe==1.5.0
requests>=2.31.0
pynacl>=1.5.0
pycryptodome>=3.19.0
opencv-python>=4.8.0
customtkinter>=5.1.3
flask>=2.3.0
Flask-Limiter>=3.5.0
numpy>=1.24.0
Pillow
face-recognition
cmake
dlib
ttkbootstrap>=1.10.1